requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
configparser==5.0.0
future==0.18.2
grpcio==1.28.1
idna==2.9
jsonschema==2.6.0
protobuf==3.11.3
pycparser==2.20
python-dotenv==0.13.0
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
